lots of refactor
