+++ internal/ansible/conf.go
@@ -1,14 +1,49 @@
 package ansible
 
 import (
+	"encoding/json"
+	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 
+	"github.com/OpenPaas/openpaas/internal/conf"
 	"gopkg.in/yaml.v3"
 )
 
 type Inventory struct {
 	All All `yaml:"all"`
+}
+
+type InventoryJson struct {
+	Clients              HostValues `json:"client_servers"`
+	NomadServers         HostValues `json:"nomad_servers"`
+	ObservabilityServers HostValues `json:"o11y_servers"`
+	VaultServers         HostValues `json:"vault_servers"`
+	ConsulServers        HostValues `json:"consul_servers"`
+	ConsulVolumes        Volumes    `json:"consul_volumes"`
+	ClientVolumes        Volumes    `json:"client_volumes"`
+}
+type Volumes struct {
+	Value []Volume `json:"value"`
+}
+
+type Volume struct {
+	Mount    string `json:"mount"`
+	Name     string `json:"name"`
+	Path     string `json:"path"`
+	ServerID int    `json:"server_id"`
+}
+
+type HostValues struct {
+	Value []Host `json:"value"`
+}
+
+type Host struct {
+	Host      string `json:"host"`
+	HostName  string `json:"host_name"`
+	PrivateIP string `json:"private_ip"`
+	ServerID  string `json:"server_id"`
 }
 
 type All struct {
@@ -111,4 +146,165 @@ func LoadInventory(file string) (*Inventory, error) {
 		return nil, err
 	}
 	return &config, nil
+}
+
+func GenerateInventory(config *conf.Config) error {
+	jsonFile, err := os.Open(filepath.Clean(filepath.Join(config.BaseDir, "inventory-output.json")))
+	if err != nil {
+		return err
+	}
+	defer func() {
+		e := jsonFile.Close()
+		fmt.Println(e)
+	}()
+	byteValue, err := io.ReadAll(jsonFile)
+	if err != nil {
+		return err
+	}
+	var inventory InventoryJson
+
+	err = json.Unmarshal(byteValue, &inventory)
+	if err != nil {
+		return err
+	}
+
+	inv := Inventory{
+		All: All{
+			Children: Children{
+				Clients:       HostGroup{Hosts: make(map[string]AnsibleHost)},
+				NomadServers:  HostGroup{Hosts: make(map[string]AnsibleHost)},
+				ConsulServers: HostGroup{Hosts: make(map[string]AnsibleHost)},
+				VaultServers:  HostGroup{Hosts: make(map[string]AnsibleHost)},
+				Grafana:       HostGroup{Hosts: make(map[string]AnsibleHost)},
+				Prometheus:    HostGroup{Hosts: make(map[string]AnsibleHost)},
+				Loki:          HostGroup{Hosts: make(map[string]AnsibleHost)},
+				Tempo:         HostGroup{Hosts: make(map[string]AnsibleHost)},
+			},
+		},
+	}
+
+	consulHosts := inventory.ConsulServers.Value
+	if len(consulHosts) == 0 {
+		consulHosts = inventory.NomadServers.Value
+	}
+
+	for _, v := range consulHosts {
+		fmt.Println(v)
+		found := false
+		for _, vol := range inventory.ConsulVolumes.Value {
+			if fmt.Sprintf("%v", vol.ServerID) == v.ServerID {
+				found = true
+				inv.All.Children.ConsulServers.Hosts[v.Host] = AnsibleHost{
+					PrivateIP: v.PrivateIP,
+					HostName:  v.HostName,
+					Mounts: []Mount{
+						{
+							Name:      "consul",
+							Path:      "/opt/consul",
+							MountPath: vol.Mount,
+							IsNomad:   false,
+							Owner:     "consul",
+						},
+					},
+					ExtraVars: map[string]string{
+						"bootstrap_expect": fmt.Sprintf("%v", len(consulHosts)),
+						"datacenter":       config.DC,
+					},
+				}
+			}
+		}
+		if !found {
+			inv.All.Children.ConsulServers.Hosts[v.Host] = AnsibleHost{
+				PrivateIP: v.PrivateIP,
+				HostName:  v.HostName,
+			}
+		}
+	}
+
+	for _, v := range inventory.NomadServers.Value {
+		inv.All.Children.NomadServers.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+			ExtraVars: map[string]string{
+				"bootstrap_expect": fmt.Sprintf("%v", len(inventory.NomadServers.Value)),
+				"datacenter":       config.DC,
+			},
+		}
+	}
+
+	for _, v := range inventory.VaultServers.Value {
+		inv.All.Children.VaultServers.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+	}
+
+	for _, v := range inventory.Clients.Value {
+		mounts := []Mount{}
+		for _, vol := range inventory.ClientVolumes.Value {
+			if fmt.Sprintf("%v", vol.ServerID) == v.ServerID {
+				mounts = append(mounts, Mount{
+					Name:      vol.Name,
+					Path:      vol.Path,
+					MountPath: vol.Mount,
+					IsNomad:   true,
+					Owner:     config.CloudProviderConfig.User,
+				})
+			}
+		}
+		inv.All.Children.Clients.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+			Mounts:    mounts,
+		}
+
+	}
+
+	if len(inventory.ObservabilityServers.Value) == 1 {
+		v := inventory.ObservabilityServers.Value[0]
+		inv.All.Children.Grafana.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+		inv.All.Children.Prometheus.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+		inv.All.Children.Loki.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+		inv.All.Children.Tempo.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+	} else {
+		v := inventory.ObservabilityServers.Value[0]
+		inv.All.Children.Grafana.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+		v = inventory.ObservabilityServers.Value[1]
+		inv.All.Children.Prometheus.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+		v = inventory.ObservabilityServers.Value[2]
+		inv.All.Children.Loki.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+		v = inventory.ObservabilityServers.Value[3]
+		inv.All.Children.Tempo.Hosts[v.Host] = AnsibleHost{
+			PrivateIP: v.PrivateIP,
+			HostName:  v.HostName,
+		}
+	}
+
+	bytes, err := yaml.Marshal(inv)
+	if err != nil {
+		return err
+	}
+
+	return os.WriteFile(filepath.Clean(filepath.Join(config.BaseDir, "inventory")), bytes, 0600)
 }
