+++ internal/hashistack/init_terraform.go
@@ -1,4 +1,4 @@
-package internal
+package hashistack
 
 import (
 	"context"
