+++ internal/hashistack/init_terraform_test.go
@@ -1,4 +1,4 @@
-package internal
+package hashistack
 
 import (
 	"context"
