+++ internal/hashistack/templates/terraform/hetzner/main.tf
