+++ internal/hashistack/templates/terraform/hetzner/vars.tf
